% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babies}
\alias{babies}
\title{Low birth weight births}
\format{
A tibble with 100 rows and 10 variables:
\describe{
  \item{\code{headcirc}}{Head circumferencem, in cm.}
  \item{\code{length}}{Birth length, in cm.}
  \item{\code{gestage}}{Gestational age in weeks.}
  \item{\code{birthwt}}{Birth weight, in grams.}
  \item{\code{momage}}{Age of mother, at birth.}
  \item{\code{toxemia}}{1 if child's mother was diagnosed with toxemia
  during pregnancy, 0 otherwise.}
  \item{\code{sbp}}{Systolic blood pressure measurement for the infant.}
  \item{\code{sex}}{Sex at birth where 1 is male and 0 is female.}
  \item{\code{grmhrm}}{1 if the infant experienced a
  germinal matrix hemorrhage, 0 otherwise.}
  \item{\code{apgar5}}{Infant's 5 minute apgar score.}
}
}
\source{
{Leviton, A., Fenton, T., Kuban, K.C. and Pagano, M., 1991.
Labor and deliver characteristics and the risk of germinal matrix
hemorrhage in low birth weight infants. Journal of child neurology,
6(1), pp.35-40 VIA Pagano, M., & Gauvreau, K. (2018).
 Principles of biostatistics. CRC Press.}
}
\usage{
babies
}
\description{
{Information from a sample of 100 low birth weight
 infants born in two teaching hospitals in Boston, Massachusetts.}
}
\keyword{datasets}
