% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bed}
\alias{bed}
\title{{Community hospital bed data}}
\format{
A tibble with 51 rows and 3 variables:
\describe{
  \item{\code{bed80}}{The number of community hospital beds per 1000 people in 1980.}
  \item{\code{bed86}}{The number of community hospital beds per 1000 people in 1986.}
  \item{\code{state}}{State}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
bed
}
\description{
{The number of community hospital beds per 1000 in each
 state in the United States and the District of Columbia.}
}
\keyword{datasets}
