% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unicef}
\alias{unicef}
\title{DATASET_TITLE}
\format{
A data frame with 144 rows and 4 variables:
\describe{
  \item{\code{nation}}{character COLUMN_DESCRIPTION}
  \item{\code{lowbwt}}{double COLUMN_DESCRIPTION}
  \item{\code{life60}}{double COLUMN_DESCRIPTION}
  \item{\code{life92}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
unicef
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
