% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{smoking}
\format{
A data frame with 1691 rows and 12 variables:
\describe{
  \item{\code{gender}}{character COLUMN_DESCRIPTION}
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{maritalstatus}}{character COLUMN_DESCRIPTION}
  \item{\code{highestqualification}}{character COLUMN_DESCRIPTION}
  \item{\code{nationality}}{character COLUMN_DESCRIPTION}
  \item{\code{ethnicity}}{character COLUMN_DESCRIPTION}
  \item{\code{grossincome}}{character COLUMN_DESCRIPTION}
  \item{\code{region}}{character COLUMN_DESCRIPTION}
  \item{\code{smoke}}{character COLUMN_DESCRIPTION}
  \item{\code{amtweekends}}{double COLUMN_DESCRIPTION}
  \item{\code{amtweekdays}}{double COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
smoking
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
