% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unicef}
\alias{unicef}
\title{Low birth weight rates worldwide}
\format{
A tibble with 144 rows and 4 variables:
\describe{
  \item{\code{nation}}{Country}
  \item{\code{lowbwt}}{Low birth-weight rates per 100 births for the year 1991}
  \item{\code{life60}}{Life expectancy at birth for the year 1960}
  \item{\code{life92}}{Life expectancy at birth for the year 1992}
}
}
\source{
{United Nations Children's Fund, 1991 via Pagano, M., &
 Gauvreau, K. (2018). Principles of biostatistics. CRC Press.,
  1993, p. 55}
}
\usage{
unicef
}
\description{
{Data collected from United Nations Children's Fund, 1991 via Pagano and Gauvreau, 1993.
 These data describe the relationship between low birth weight
  births and life expectancy at the country level.}
}
\keyword{datasets}
